// week-3/enhanced-composer-app/src/app/composer.interface.ts
/**
==========================================
  Title: enhanced composer app
  Author: Professor Krasso
  Modifier: Sarah Jean Baptiste
  Date: 08/19/2021
  Description: enhanced-composer-app
==========================================
*/

// export statement
export interface IComposer {
  composerId: number;
  fullName: string;
  genre: string;
}